Hack for emmc with noncontiguous partition numbers
On the Samsung Galaxy 5.0, all partitions are on the internal sdcard
(eMMC). The first partition is defined as an extended partion, with no
other primary partitions used.  This results in the partitions showing
up as p1,p5,p6,p7,..,p17 with no p2,p3,p4.  In DirectVolume, it doesn't
set the disk -> State_Idle until it sees all partitions.  And not
knowing about this extended partition issue, vold assumes that 1->MAX
partitions all exist, so it waits forever for p2,p3,p4 to show up.

Not sure what the proper fix is to determine whether "all" partitions
show up, but this hack causes the disk to go idle when the one we care
about (the one partition defined in the vold.fstab for this device)
shows up.  Works for me..

Change-Id: I74b223e782a4c661678f6d1bffb7eda50370aeda